Script in Docker implementing the VNF is explicitly started by the un-orchestrator.
Each container must include, in the root foder, the script start.sh.
This script is then run by the un-orchestrator after that it starts the container
and pushes in it all the proper interfaces.

The it is no longer necessary that the script starts with a pause and terminates
with a infinite loop.

Moreover, the container image should not have a script executed automatically
when the containers starts.